Update: more code reorganization.
