+++ droneyourfood/lib/Authentication/Authentication.dart
@@ -1,17 +1,15 @@
-import 'package:droneyourfood/Shopping/Shopping.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:google_sign_in/google_sign_in.dart';
 
-import 'package:droneyourfood/Components/ScrollColumn.dart';
 import 'package:droneyourfood/main.dart';
 import 'package:droneyourfood/Tools.dart';
+import 'package:droneyourfood/Components/ScrollColumn.dart';
+import 'package:droneyourfood/Shopping/Shopping.dart';
 
 abstract class AuthState<T extends StatefulWidget> extends State<T> {
-  TextEditingController _emailField = TextEditingController();
-  TextEditingController _passwordField = TextEditingController();
-  String _error = "";
+  TextEditingController emailField = TextEditingController();
+  TextEditingController passwordField = TextEditingController();
+  String error = "";
 
   List<Widget> genButtons(BuildContext context, final double fieldWidth);
 
@@ -44,11 +42,11 @@ abstract class AuthState<T extends StatefulWidget> extends State<T> {
     return [
       SizedBox(
         width: fieldWidth,
-        child: genInputField(context, _emailField, "Email", false),
+        child: genInputField(context, emailField, "Email", false),
       ),
       SizedBox(
         width: fieldWidth,
-        child: genInputField(context, _passwordField, "Password", true),
+        child: genInputField(context, passwordField, "Password", true),
       ),
     ];
   }
@@ -56,294 +54,8 @@ abstract class AuthState<T extends StatefulWidget> extends State<T> {
   Widget genError(BuildContext context) {
     /* ERROR DISPLAYING AFTER FAILURE */
     return Container(
-      child: Text(this._error,
+      child: Text(this.error,
           textAlign: TextAlign.center, style: TextStyle(color: Colors.red)),
     );
-  }
-}
-
-class SignIn extends StatefulWidget {
-  @override
-  _SignInState createState() => _SignInState();
-}
-
-class _SignInState extends AuthState<SignIn> {
-  void navigateToRegisterScreen(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => Register()),
-    );
-  }
-
-  void navigateToPasswordResetScreen(BuildContext context) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => ResetPassword()),
-    );
-  }
-
-  void signInWithGoogle() async {
-    // Trigger the authentication flow
-    final GoogleSignInAccount googleUser = await GoogleSignIn().signIn();
-    // Obtain the auth details from the request
-    final GoogleSignInAuthentication googleAuth =
-        await googleUser.authentication;
-    // Create a new credential
-    final GoogleAuthCredential credential = GoogleAuthProvider.credential(
-      accessToken: googleAuth.accessToken,
-      idToken: googleAuth.idToken,
-    );
-
-    // Once signed in, return the UserCredential
-    await FirebaseAuth.instance.signInWithCredential(credential);
-    // fogao -> >:(
-    navigateToHomeScreen(context);
-  }
-
-  void login() async {
-    try {
-      UserCredential userCredential = await FirebaseAuth.instance
-          .signInWithEmailAndPassword(
-              email: _emailField.text, password: _passwordField.text);
-
-      if (!userCredential.user.emailVerified) {
-        setState(() {
-          this._error = "Email not verified. Verification email sent";
-        });
-        await userCredential.user.sendEmailVerification();
-      } else {
-        // Only works if the user is verified
-        navigateToHomeScreen(context);
-      }
-    } on FirebaseAuthException catch (e) {
-      setState(() {
-        if (e.code == 'user-not-found')
-          this._error = "No user found for that email.";
-        else if (e.code == 'wrong-password')
-          this._error = "Wrong password provided for that user.";
-        else
-          this._error = e.message;
-      });
-    }
-  }
-
-  List<Widget> genButtons(BuildContext context, final double fieldWidth) {
-    final double fontSize = 16;
-
-    return [
-      /* LOGIN BUTTON */
-      ElevatedButton(
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all<Color>(Colors.blue),
-          minimumSize:
-              MaterialStateProperty.all<Size>(Size(fieldWidth, fontSize * 2)),
-        ),
-        child: Text("Log in", style: TextStyle(fontSize: fontSize)),
-        onPressed: login,
-      ),
-      ElevatedButton(
-          style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all<Color>(Colors.red),
-            minimumSize:
-                MaterialStateProperty.all<Size>(Size(fieldWidth, fontSize * 2)),
-          ),
-          child:
-              Text("Sign In with Google", style: TextStyle(fontSize: fontSize)),
-          onPressed: signInWithGoogle),
-      /* REGISTER BUTTON */
-      ElevatedButton(
-        style: ButtonStyle(
-          minimumSize:
-              MaterialStateProperty.all<Size>(Size(fieldWidth, fontSize * 2)),
-        ),
-        child: Text("Sign Up", style: TextStyle(fontSize: fontSize)),
-        onPressed: () {
-          navigateToRegisterScreen(context);
-        },
-      ),
-      RichText(
-        text: TextSpan(
-          style: TextStyle(color: Colors.grey),
-          text: "Forgot your password?",
-          recognizer: TapGestureRecognizer()
-            ..onTap = () {
-              navigateToPasswordResetScreen(context);
-            },
-        ),
-      ),
-      genError(context)
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final double fieldWidth = MediaQuery.of(context).size.width * 0.8;
-
-    return Scaffold(
-      appBar: AppBar(title: Text("Drone your food - Log In")),
-      body: Center(
-        child: ScrollColumn(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children:
-              genInputs(context, fieldWidth) + genButtons(context, fieldWidth),
-        ),
-      ),
-    );
-  }
-}
-
-class Register extends StatefulWidget {
-  @override
-  _RegisterState createState() => _RegisterState();
-}
-
-class _RegisterState extends AuthState<Register> {
-  TextEditingController _userNameField = TextEditingController();
-
-  void register() async {
-    try {
-      UserCredential userCredential = await FirebaseAuth.instance
-          .createUserWithEmailAndPassword(
-              email: _emailField.text, password: _passwordField.text);
-      await userCredential.user.updateProfile(displayName: _userNameField.text);
-
-      setState(() {
-        this._error = "Verification email sent";
-      });
-      await userCredential.user.sendEmailVerification();
-    } on FirebaseAuthException catch (e) {
-      setState(() {
-        if (e.code == 'weak-password')
-          this._error = "The password provided is too weak.";
-        else if (e.code == 'email-already-in-use')
-          this._error = "An account already exists for that email.";
-        else
-          this._error = e.message;
-      });
-    }
-  }
-
-  @override
-  List<Widget> genInputs(BuildContext context, final double fieldWidth) {
-    return <Widget>[
-          SizedBox(
-            width: fieldWidth,
-            child: genInputField(context, _userNameField, "Username", false),
-          )
-        ] +
-        super.genInputs(context, fieldWidth);
-  }
-
-  List<Widget> genButtons(BuildContext context, final double fieldWidth) {
-    final double fontSize = 16;
-
-    /*
-     * !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-     * ! FOI A ANA QUE ESCOLHEU O ROSA !
-     * ! XAU                           !
-     * !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-     */
-
-    return [
-      /* REGISTER BUTTON */
-      ElevatedButton(
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all<Color>(Colors.pink),
-          minimumSize:
-              MaterialStateProperty.all<Size>(Size(fieldWidth, fontSize * 2)),
-        ),
-        child: Text("Sign Up", style: TextStyle(fontSize: fontSize)),
-        onPressed: register,
-      ),
-      genError(context)
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final double fieldWidth = MediaQuery.of(context).size.width * 0.8;
-
-    return Scaffold(
-      appBar: AppBar(title: Text("Drone your food - Sign Up")),
-      body: Center(
-        child: ScrollColumn(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children:
-              genInputs(context, fieldWidth) + genButtons(context, fieldWidth),
-        ),
-      ),
-    );
-  }
-}
-
-class ResetPassword extends StatefulWidget {
-  @override
-  _ResetPasswordState createState() => _ResetPasswordState();
-}
-
-class _ResetPasswordState extends AuthState<ResetPassword> {
-  TextEditingController _emailField = TextEditingController();
-
-  void resetPassword() async {
-    //Error reset
-    this._error = "";
-    try {
-      await FirebaseAuth.instance
-          .sendPasswordResetEmail(email: _emailField.text);
-      setState(() {
-        this._error = "Email sent";
-      });
-    } on FirebaseAuthException catch (e) {
-      setState(() {
-        if (e.code == "user-not-found") {
-          this._error = "Email sent";
-        } else {
-          this._error = e.message;
-        }
-      });
-    }
-  }
-
-  @override
-  List<Widget> genInputs(BuildContext context, final double fieldWidth) {
-    return <Widget>[
-      SizedBox(
-        width: fieldWidth,
-        child: genInputField(context, _emailField, "Email", false),
-      )
-    ];
-  }
-
-  List<Widget> genButtons(BuildContext context, final double fieldWidth) {
-    final double fontSize = 16;
-
-    return [
-      /* REGISTER BUTTON */
-      ElevatedButton(
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.all<Color>(Colors.orange),
-          minimumSize:
-              MaterialStateProperty.all<Size>(Size(fieldWidth, fontSize * 2)),
-        ),
-        child: Text("Send password reset email",
-            style: TextStyle(fontSize: fontSize)),
-        onPressed: resetPassword,
-      ),
-      genError(context)
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final double fieldWidth = MediaQuery.of(context).size.width * 0.8;
-
-    return Scaffold(
-        appBar: AppBar(title: Text("Drone your food - Reset Password")),
-        body: Center(
-            child: ScrollColumn(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children:
-              genInputs(context, fieldWidth) + genButtons(context, fieldWidth),
-        )));
   }
 }
